+++ .github/CHANGELOG.md
@@ -52,6 +52,7 @@
 1. [ECAM] Revised fuel page to match neo - @RichardPilbery (tricky_dicky#3571)
 1. [RMP] Added 25 kHz VHF spacing option - @tyler58546 (tyler58546)
 1. [MCDU] Scratchpad rewrite - @derl30n (Leon)
+1. [ATHR/ENGINE] Improved CLB thrust limit for improved FLEX N1 numbers - @aguther (Andreas Guther)
 
 ## 0.7.0